fix codec check

+++ services/media_engine/plugins/ffmpeg_adapter/muxer/mpeg4_muxer/avc_parser.cpp
@@ -102,8 +102,8 @@ int32_t AvcParser::ParseSps(const uint8_t* sample, int32_t size)
     }
 
     std::vector<uint8_t> sps(0x02, 0);
-    sps[1] = static_cast<uint8_t>(size & 0xff);
-    sps[0] = static_cast<uint8_t>((size >> 0x08) & 0xff);
+    sps[1] = static_cast<uint8_t>(static_cast<uint32_t>(size) & 0xff);
+    sps[0] = static_cast<uint8_t>((static_cast<uint32_t>(size) >> 0x08) & 0xff);
     sps.insert(sps.end(), sample, sample + size);
     avccBox_->sps_.emplace_back(sps);
     ++spsCount_;
@@ -123,8 +123,8 @@ int32_t AvcParser::ParsePps(const uint8_t* sample, int32_t size)
     }
 
     std::vector<uint8_t> pps(2, 0);  // 2
-    pps[1] = static_cast<uint8_t>(size & 0xff);
-    pps[0] = static_cast<uint8_t>((size >> 0x08) & 0xff);
+    pps[1] = static_cast<uint8_t>(static_cast<uint32_t>(size) & 0xff);
+    pps[0] = static_cast<uint8_t>((static_cast<uint32_t>(size) >> 0x08) & 0xff);
     pps.insert(pps.end(), sample, sample + size);
     avccBox_->pps_.emplace_back(pps);
     ++avccBox_->ppsCount_;
@@ -143,8 +143,8 @@ int32_t AvcParser::ParseSpsExt(const uint8_t* sample, int32_t size)
     }
 
     std::vector<uint8_t> spsExt(2, 0);  // 2
-    spsExt[1] = static_cast<uint8_t>(size & 0xff);
-    spsExt[0] = static_cast<uint8_t>((size >> 0x08) & 0xff);
+    spsExt[1] = static_cast<uint8_t>(static_cast<uint32_t>(size) & 0xff);
+    spsExt[0] = static_cast<uint8_t>((static_cast<uint32_t>(size) >> 0x08) & 0xff);
     spsExt.insert(spsExt.end(), sample, sample + size);
     avccBox_->spsExt_.emplace_back(spsExt);
     ++avccBox_->spsExtCount_;
